Extract StringReplacementEvaluator to its own file
